update openclah & passwall

+++ applications/luci-app-openclash/root/usr/share/openclash/openclash_rule.sh
@@ -3,19 +3,24 @@
 . /lib/functions.sh
 . /usr/share/openclash/ruby.sh
 . /usr/share/openclash/log.sh
+. /usr/share/openclash/openclash_curl.sh
+. /usr/share/openclash/uci.sh
 
-   set_lock() {
-      exec 877>"/tmp/lock/openclash_rule.lock" 2>/dev/null
-      flock -x 877 2>/dev/null
-   }
+set_lock() {
+   exec 877>"/tmp/lock/openclash_rule.lock" 2>/dev/null
+   flock -x 877 2>/dev/null
+}
 
-   del_lock() {
-      flock -u 877 2>/dev/null
-      rm -rf "/tmp/lock/openclash_rule.lock"
-   }
+del_lock() {
+   flock -u 877 2>/dev/null
+   rm -rf "/tmp/lock/openclash_rule.lock" 2>/dev/null
+}
 
-   yml_other_rules_dl()
-   {
+set_lock
+inc_job_counter
+
+yml_other_rules_dl()
+{
    local section="$1"
    local enabled config
    config_get_bool "enabled" "$section" "enabled" "1"
@@ -36,51 +41,24 @@
    if [ "$rule_name" = "lhie1" ]; then
       if [ "$github_address_mod" != "0" ]; then
          if [ "$github_address_mod" == "https://cdn.jsdelivr.net/" ] || [ "$github_address_mod" == "https://fastly.jsdelivr.net/" ] || [ "$github_address_mod" == "https://testingcf.jsdelivr.net/" ]; then
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 "$github_address_mod"gh/dler-io/Rules@master/Clash/Rule.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-         elif [ "$github_address_mod" == "https://raw.fastgit.org/" ]; then
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 https://raw.fastgit.org/dler-io/Rules/master/Clash/Rule.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
+            DOWNLOAD_URL="${github_address_mod}gh/dler-io/Rules@master/Clash/Rule.yaml"
          else
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 "$github_address_mod"https://raw.githubusercontent.com/dler-io/Rules/master/Clash/Rule.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
+            DOWNLOAD_URL="${github_address_mod}https://raw.githubusercontent.com/dler-io/Rules/master/Clash/Rule.yaml"
          fi
       else
-         curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 https://raw.githubusercontent.com/dler-io/Rules/master/Clash/Rule.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-      fi
-      sed -i '1i rules:' /tmp/rules.yaml
-   elif [ "$rule_name" = "ConnersHua" ]; then
-      if [ "$github_address_mod" != "0" ]; then
-         if [ "$github_address_mod" == "https://cdn.jsdelivr.net/" ] || [ "$github_address_mod" == "https://fastly.jsdelivr.net/" ] || [ "$github_address_mod" == "https://testingcf.jsdelivr.net/" ]; then
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 "$github_address_mod"gh/DivineEngine/Profiles@master/Clash/Outbound.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-         elif [ "$github_address_mod" == "https://raw.fastgit.org/" ]; then
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 https://raw.fastgit.org/DivineEngine/Profiles/master/Clash/Outbound.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-         else
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 "$github_address_mod"https://raw.githubusercontent.com/DivineEngine/Profiles/master/Clash/Outbound.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-         fi
-      else
-         curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 https://raw.githubusercontent.com/DivineEngine/Profiles/master/Clash/Outbound.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-      fi
-      sed -i "s/# - RULE-SET,ChinaIP,DIRECT/- RULE-SET,ChinaIP,DIRECT/g" /tmp/rules.yaml 2>/dev/null
-      sed -i "s/- GEOIP,/#- GEOIP,/g" /tmp/rules.yaml 2>/dev/null
-   elif [ "$rule_name" = "ConnersHua_return" ]; then
-      if [ "$github_address_mod" != "0" ]; then
-         if [ "$github_address_mod" == "https://cdn.jsdelivr.net/" ] || [ "$github_address_mod" == "https://fastly.jsdelivr.net/" ] || [ "$github_address_mod" == "https://testingcf.jsdelivr.net/" ]; then
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 "$github_address_mod"gh/DivineEngine/Profiles@master/Clash/Inbound.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-         elif [ "$github_address_mod" == "https://raw.fastgit.org/" ]; then
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 https://raw.fastgit.org/DivineEngine/Profiles/master/Clash/Inbound.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-         else
-            curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 "$github_address_mod"https://raw.githubusercontent.com/DivineEngine/Profiles/master/Clash/Inbound.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
-         fi
-      else
-         curl -SsL --connect-timeout 30 -m 60 --speed-time 30 --speed-limit 1 --retry 2 https://raw.githubusercontent.com/DivineEngine/Profiles/master/Clash/Inbound.yaml -o /tmp/rules.yaml 2>&1 | awk -v time="$(date "+%Y-%m-%d %H:%M:%S")" -v file="/tmp/rules.yaml" '{print time "【" file "】Download Failed:【"$0"】"}' >> "$LOG_FILE"
+         DOWNLOAD_URL="https://raw.githubusercontent.com/dler-io/Rules/master/Clash/Rule.yaml"
       fi
    fi
-   if [ -s "/tmp/rules.yaml" ]; then
+   DOWNLOAD_FILE_CURL "$DOWNLOAD_URL" "/tmp/rules.yaml"
+   if [ "$?" -eq 0 ] && [ -s "/tmp/rules.yaml" ]; then
       LOG_OUT "Download Successful, Start Preprocessing Rule File..."
+      sed -i '1i rules:' /tmp/rules.yaml
       ruby -ryaml -rYAML -I "/usr/share/openclash" -E UTF-8 -e "
       begin
       YAML.load_file('/tmp/rules.yaml');
       rescue Exception => e
-      puts '${LOGTIME} Error: Unable To Parse Updated Rules File,【${rule_name}:' + e.message + '】'
-      system 'rm -rf /tmp/rules.yaml 2>/dev/null'
+         YAML.LOG('Error: Unable To Parse Updated Rules File,【${rule_name}:' + e.message + '】');
+         system 'rm -rf /tmp/rules.yaml 2>/dev/null';
       end
       " 2>/dev/null >> $LOG_FILE
       if [ $? -ne 0 ]; then
@@ -105,9 +83,17 @@
       elif ! "$(ruby -ryaml -rYAML -I "/usr/share/openclash" -E UTF-8 -e "
          Value = YAML.load_file('/usr/share/openclash/res/${rule_name}.yaml');
          Value_1 = YAML.load_file('/tmp/rules.yaml');
-         OLD_GROUP = Value['rules'].collect{|x| x.split(',')[2] or x.split(',')[1]}.uniq;
-         NEW_GROUP = Value_1['rules'].collect{|x| x.split(',')[2] or x.split(',')[1]}.uniq;
-         puts (OLD_GROUP | NEW_GROUP).eql?(OLD_GROUP)
+         OLD_GROUP = Value['rules'].collect{|x| x.split(',')[2] or x.split(',')[1]}.uniq.map(&:strip);
+         NEW_GROUP = Value_1['rules'].collect{|x| x.split(',')[2] or x.split(',')[1]}.uniq.map(&:strip);
+         if (OLD_GROUP | NEW_GROUP).eql?(OLD_GROUP) then
+            if (OLD_GROUP | NEW_GROUP).eql?(NEW_GROUP) then
+               puts true
+            else
+               puts false
+            end
+         else
+            puts false
+         end
          ")" && [ -f "/usr/share/openclash/res/${rule_name}.yaml" ]; then
          LOG_OUT "Error: Updated Others Rules【$rule_name】Has Incompatible Proxy-Group, Update Exit, Please Wait For OpenClash Update To Adapt..."
          rm -rf /tmp/rules.yaml >/dev/null 2>&1
@@ -117,7 +103,7 @@
       fi
       
       #取出规则部分
-      ruby_read "/tmp/rules.yaml" ".select {|x| 'rule-providers' == x or 'script' == x or 'rules' == x }.to_yaml" > "$OTHER_RULE_FILE"
+      ruby_read "/tmp/rules.yaml" ".select {|x| 'rule-providers' == x or 'rules' == x }.to_yaml" > "$OTHER_RULE_FILE"
       #合并
       cat "$OTHER_RULE_FILE" > "/tmp/rules.yaml" 2>/dev/null
       rm -rf /tmp/other_rule* 2>/dev/null
@@ -128,54 +114,52 @@
          LOG_OUT "Detected that The Downloaded Rule File Has Been Updated, Starting To Replace..."
          mv /tmp/rules.yaml /usr/share/openclash/res/"$rule_name".yaml >/dev/null 2>&1
          LOG_OUT "Other Rules【$rule_name】Update Successful!"
-         ifrestart=1
+         restart=1
       else
          LOG_OUT "Updated Other Rules【$rule_name】No Change, Do Nothing!"
       fi
    else
       LOG_OUT "Other Rules【$rule_name】Update Error, Please Try Again Later..."
    fi
-   }
-   
-   LOGTIME=$(echo $(date "+%Y-%m-%d %H:%M:%S"))
-   LOG_FILE="/tmp/openclash.log"
-   RUlE_SOURCE=$(uci get openclash.config.rule_source 2>/dev/null)
-   github_address_mod=$(uci -q get openclash.config.github_address_mod || echo 0)
-   set_lock
-   
-   if [ "$RUlE_SOURCE" = "0" ]; then
-      LOG_OUT "Other Rules Not Enable, Update Stop!"
-   else
-      OTHER_RULE_FILE="/tmp/other_rule.yaml"
-      CONFIG_FILE=$(uci get openclash.config.config_path 2>/dev/null)
-      CONFIG_NAME=$(echo "$CONFIG_FILE" |awk -F '/' '{print $5}' 2>/dev/null)
-      ifrestart=0
-   
-      if [ -z "$CONFIG_FILE" ]; then
-         for file_name in /etc/openclash/config/*
-         do
-            if [ -f "$file_name" ]; then
-               CONFIG_FILE=$file_name
-               CONFIG_NAME=$(echo "$CONFIG_FILE" |awk -F '/' '{print $5}' 2>/dev/null)
-               break
-            fi
-         done
-      fi
+}
 
-      if [ -z "$CONFIG_NAME" ]; then
-         CONFIG_FILE="/etc/openclash/config/config.yaml"
-         CONFIG_NAME="config.yaml"
-      fi
-      
-      config_load "openclash"
-      config_foreach yml_other_rules_dl "other_rules" "$CONFIG_NAME"
-      if [ -z "$rule_name" ]; then
-        LOG_OUT "Get Other Rules Settings Faild, Update Stop!"
-      fi
-      if [ "$ifrestart" -eq 1 ] && [ "$(unify_ps_prevent)" -eq 0 ] && [ "$(find /tmp/lock/ |grep -v "openclash.lock" |grep -c "openclash")" -le 1 ]; then
-         /etc/init.d/openclash restart >/dev/null 2>&1 &
-      fi
+LOG_FILE="/tmp/openclash.log"
+RUlE_SOURCE=$(uci_get_config "rule_source")
+github_address_mod=$(uci_get_config "github_address_mod" || echo 0)
+restart=0
+
+if [ "$RUlE_SOURCE" = "0" ]; then
+   LOG_OUT "Other Rules Not Enable, Update Stop!"
+else
+   OTHER_RULE_FILE="/tmp/other_rule.yaml"
+   CONFIG_FILE=$(uci_get_config "config_path")
+   CONFIG_NAME=$(echo "$CONFIG_FILE" |awk -F '/' '{print $5}' 2>/dev/null)
+
+   if [ -z "$CONFIG_FILE" ]; then
+      for file_name in /etc/openclash/config/*
+      do
+         if [ -f "$file_name" ]; then
+            CONFIG_FILE=$file_name
+            CONFIG_NAME=$(echo "$CONFIG_FILE" |awk -F '/' '{print $5}' 2>/dev/null)
+            break
+         fi
+      done
    fi
-   rm -rf /tmp/rules.yaml >/dev/null 2>&1
-   SLOG_CLEAN
-   del_lock
+
+   if [ -z "$CONFIG_NAME" ]; then
+      CONFIG_FILE="/etc/openclash/config/config.yaml"
+      CONFIG_NAME="config.yaml"
+   fi
+   
+   config_load "openclash"
+   config_foreach yml_other_rules_dl "other_rules" "$CONFIG_NAME"
+   if [ -z "$rule_name" ]; then
+     LOG_OUT "Get Other Rules Settings Faild, Update Stop!"
+   fi
+fi
+
+rm -rf /tmp/rules.yaml >/dev/null 2>&1
+
+SLOG_CLEAN
+dec_job_counter_and_restart "$restart"
+del_lock
